requirements.txt
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
numpy>=1.26.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn>=0.29.0
rank-bm25>=0.2.2
scikit-learn>=1.4.0
transformers>=4.40.0
torch>=2.0.0
python-dotenv>=1.0.0
